+++ backend/windmill-common/src/users.rs
@@ -17,7 +17,11 @@ pub fn username_to_permissioned_as(user: &str) -> String {
 }
 
 pub fn truncate_token(token: &str) -> String {
-    let mut s = token[..10].to_owned();
-    s.push_str("*****");
-    s
+    if token.len() > 10 {
+        let mut s = token[..10].to_owned();
+        s.push_str("*****");
+        s
+    } else {
+        token.to_string()
+    }
 }