+++ frontend/src/lib/components/apps/editor/GridEditor.svelte
@@ -32,7 +32,7 @@
 		allIdsInPath
 	} = getContext<AppViewerContext>('AppViewerContext')
 
-	const { history } = getContext<AppEditorContext>('AppEditorContext')
+	const { history, scale } = getContext<AppEditorContext>('AppEditorContext')
 
 	let previousSelectedIds: string[] | undefined = undefined
 	$: if (!deepEqual(previousSelectedIds, $selectedComponent)) {
@@ -73,9 +73,9 @@
 	<div
 		style={$app.css?.['app']?.['grid']?.style}
 		class={twMerge(
-			'px-4 pt-4 pb-2 overflow-visible',
+			'p-2 overflow-visible',
 			$app.css?.['app']?.['grid']?.class ?? '',
-			'wm-app-grid'
+			'wm-app-grid border'
 		)}
 		on:pointerdown={() => {
 			$selectedComponent = undefined
@@ -83,7 +83,14 @@
 		}}
 		bind:clientWidth={$parentWidth}
 	>
-		<div class={!$focusedGrid && $mode !== 'preview' ? ' border border-dashed' : ''}>
+		<div
+			class={twMerge(
+				!$focusedGrid && $mode !== 'preview' ? 'border-dashed' : '',
+				'subgrid',
+				'border-[#999999] dark:border-[#aaaaaa] border '
+			)}
+			style={`transform: scale(${$scale / 100})`}
+		>
 			<Grid
 				allIdsInPath={$allIdsInPath}
 				selectedIds={$selectedComponent}