+++ frontend/src/routes/(root)/(logged)/+layout.svelte
@@ -34,6 +34,7 @@
 	import { SUPERADMIN_SETTINGS_HASH, USER_SETTINGS_HASH } from '$lib/components/sidebar/settings'
 	import { isCloudHosted } from '$lib/cloud'
 	import MultiplayerMenu from '$lib/components/sidebar/MultiplayerMenu.svelte'
+	import { syncTutorialsTodos } from '$lib/tutorialUtils'
 
 	OpenAPI.WITH_CREDENTIALS = true
 	let menuOpen = false
@@ -74,6 +75,7 @@
 	function onLoad() {
 		loadFavorites()
 		loadUsage()
+		syncTutorialsTodos()
 	}
 
 	async function loadUsage() {