+++ frontend/tailwind.config.cjs
@@ -583,7 +583,6 @@ const config = {
 					fontSize: '18px',
 					fontWeight: theme('fontWeight.semibold'),
 					lineHeight: '1.3',
-					color: theme('colors.gray.600'),
 					[`@media (min-width: ${theme('screens.qhd')})`]: {
 						fontSize: '20px'
 					}
@@ -592,7 +591,6 @@ const config = {
 					fontSize: '16px',
 					fontWeight: theme('fontWeight.semibold'),
 					lineHeight: '1.5',
-					color: theme('colors.gray.600'),
 					[`@media (min-width: ${theme('screens.qhd')})`]: {
 						fontSize: '18px'
 					}
@@ -601,7 +599,6 @@ const config = {
 					fontSize: '16px',
 					fontWeight: theme('fontWeight.medium'),
 					lineHeight: '1.5',
-					color: theme('colors.gray.600'),
 					[`@media (min-width: ${theme('screens.qhd')})`]: {
 						fontSize: '18px'
 					}
@@ -789,7 +786,7 @@ const config = {
 					overflow: 'auto !important'
 				},
 				'.splitpanes__splitter': {
-					backgroundColor: lightTheme.border + ' !important',
+					backgroundColor: 'rgb(243 244 246) !important',
 					margin: '0 !important',
 					border: 'none !important',
 					'&::after': {